feat(nym-sdk): gate SDK modules and dependencies behind cargo features
