+++ sdk/rust/nym-sdk/src/error.rs
@@ -1,6 +1,7 @@
 // Copyright 2025 - Nym Technologies SA <contact@nymtech.net>
 // SPDX-License-Identifier: Apache-2.0
 
+#[cfg(feature = "ipr")]
 use nym_ip_packet_requests::ConnectFailureReason;
 use nym_validator_client::nym_api::error::NymAPIError;
 use nym_validator_client::nyxd::error::NyxdError;
@@ -60,12 +61,14 @@ pub enum Error {
     #[error("socks5 channel could not be started")]
     Socks5NotStarted,
 
+    #[cfg(feature = "bandwidth")]
     #[error("bandwidth controller error: {0}")]
     BandwidthControllerError(#[from] nym_bandwidth_controller::error::BandwidthControllerError),
 
     #[error("invalid voucher blob")]
     InvalidVoucherBlob,
 
+    #[cfg(feature = "bandwidth")]
     #[error("invalid mnemonic: {0}")]
     InvalidMnemonic(#[from] bip39::Error),
 
@@ -93,6 +96,7 @@ pub enum Error {
     #[error("this operation is currently unsupported: {details}")]
     Unsupported { details: String },
 
+    #[cfg(feature = "tcp-proxy")]
     #[error(transparent)]
     Bincode(#[from] bincode::Error),
 
@@ -120,6 +124,7 @@ pub enum Error {
     #[error("unexpected response to connect request: {0}")]
     UnexpectedResponseType(String),
 
+    #[cfg(feature = "ipr")]
     #[error("connect denied: {0:?}")]
     ConnectDenied(ConnectFailureReason),
