+++ sdk/rust/nym-sdk/src/mixnet/client.rs
@@ -1,22 +1,29 @@
 // Copyright 2022-2023 - Nym Technologies SA <contact@nymtech.net>
 // SPDX-License-Identifier: Apache-2.0
 
-use super::{connection_state::BuilderState, Config, StoragePaths};
+#[cfg(feature = "fs-storage")]
+use super::StoragePaths;
+use super::{connection_state::BuilderState, Config};
+#[cfg(feature = "bandwidth")]
 use crate::bandwidth::{BandwidthAcquireClient, BandwidthImporter};
+#[cfg(feature = "socks5")]
 use crate::mixnet::socks5_client::Socks5MixnetClient;
 use crate::mixnet::{MixnetClient, Recipient};
 use crate::GatewayTransceiver;
 use crate::NymNetworkDetails;
 use crate::{Error, Result};
 use log::{debug, warn};
+#[cfg(feature = "bandwidth")]
 use nym_bandwidth_controller::BandwidthTicketProvider;
 use nym_client_core::client::base_client::storage::gateways_storage::GatewayRegistration;
 use nym_client_core::client::base_client::storage::helpers::{
     get_active_gateway_identity, get_all_registered_identities, has_gateway_details,
     set_active_gateway,
 };
+#[cfg(feature = "fs-storage")]
+use nym_client_core::client::base_client::storage::OnDiskPersistent;
 use nym_client_core::client::base_client::storage::{
-    Ephemeral, GatewaysDetailsStore, MixnetClientStorage, OnDiskPersistent,
+    Ephemeral, GatewaysDetailsStore, MixnetClientStorage,
 };
 use nym_client_core::client::base_client::BaseClientBuilder;
 use nym_client_core::client::base_client::{BaseClient, EventSender};
@@ -26,8 +33,10 @@ use nym_client_core::error::ClientCoreError;
 use nym_client_core::init::helpers::gateways_for_init;
 use nym_client_core::init::types::{GatewaySelectionSpecification, GatewaySetup};
 use nym_client_core::init::{refresh_gateway_published_data, setup_gateway};
+#[cfg(feature = "bandwidth")]
 use nym_credentials_interface::TicketType;
 use nym_crypto::hkdf::DerivationMaterial;
+#[cfg(feature = "socks5")]
 use nym_socks5_client_core::config::Socks5;
 use nym_task::ShutdownTracker;
 use nym_topology::provider_trait::TopologyProvider;
@@ -39,6 +48,7 @@ use std::path::PathBuf;
 #[cfg(unix)]
 use std::sync::Arc;
 use url::Url;
+#[cfg(feature = "bandwidth")]
 use zeroize::Zeroizing;
 
 /// The number of reply SURBs to include in a message by default.
@@ -67,13 +77,14 @@ pub(crate) const DEFAULT_NUMBER_OF_SURBS: u32 = 10;
 #[derive(Default)]
 pub struct MixnetClientBuilder<S: MixnetClientStorage = Ephemeral> {
     config: Config,
-    storage_paths: Option<StoragePaths>,
+    #[cfg(feature = "socks5")]
     socks5_config: Option<Socks5>,
 
     wait_for_gateway: bool,
     wait_for_initial_topology: bool,
     custom_topology_provider: Option<Box<dyn TopologyProvider + Send + Sync>>,
     custom_gateway_transceiver: Option<Box<dyn GatewayTransceiver + Send + Sync>>,
+    #[cfg(feature = "bandwidth")]
     custom_bandwidth_provider: Option<Box<dyn BandwidthTicketProvider>>,
     custom_shutdown: Option<ShutdownTracker>,
     event_tx: Option<EventSender>,
@@ -109,15 +120,17 @@ impl MixnetClientBuilder<Ephemeral> {
     }
 }
 
+#[cfg(feature = "fs-storage")]
 impl MixnetClientBuilder<OnDiskPersistent> {
     pub async fn new_with_default_storage(storage_paths: StoragePaths) -> Result<Self> {
         Ok(MixnetClientBuilder {
             config: Default::default(),
-            storage_paths: None,
+            #[cfg(feature = "socks5")]
             socks5_config: None,
             wait_for_gateway: false,
             wait_for_initial_topology: false,
             custom_topology_provider: None,
+            #[cfg(feature = "bandwidth")]
             custom_bandwidth_provider: None,
             storage: storage_paths
                 .initialise_default_persistent_storage()
@@ -148,12 +161,13 @@ where
     pub fn new_with_storage(storage: S) -> MixnetClientBuilder<S> {
         MixnetClientBuilder {
             config: Default::default(),
-            storage_paths: None,
+            #[cfg(feature = "socks5")]
             socks5_config: None,
             wait_for_gateway: false,
             wait_for_initial_topology: false,
             custom_topology_provider: None,
             custom_gateway_transceiver: None,
+            #[cfg(feature = "bandwidth")]
             custom_bandwidth_provider: None,
             custom_shutdown: None,
             event_tx: None,
@@ -176,12 +190,13 @@ where
     pub fn set_storage<T: MixnetClientStorage>(self, storage: T) -> MixnetClientBuilder<T> {
         MixnetClientBuilder {
             config: self.config,
-            storage_paths: self.storage_paths,
+            #[cfg(feature = "socks5")]
             socks5_config: self.socks5_config,
             wait_for_gateway: self.wait_for_gateway,
             wait_for_initial_topology: self.wait_for_initial_topology,
             custom_topology_provider: self.custom_topology_provider,
             custom_gateway_transceiver: self.custom_gateway_transceiver,
+            #[cfg(feature = "bandwidth")]
             custom_bandwidth_provider: self.custom_bandwidth_provider,
             custom_shutdown: self.custom_shutdown,
             event_tx: self.event_tx,
@@ -206,6 +221,7 @@ where
     }
 
     /// Change the underlying storage of this builder to use default implementation of on-disk disk_persistence.
+    #[cfg(feature = "fs-storage")]
     #[must_use]
     pub fn set_default_storage(
         self,
@@ -297,6 +313,7 @@ where
     }
 
     /// Configure the SOCKS5 mode.
+    #[cfg(feature = "socks5")]
     #[must_use]
     pub fn socks5_config(mut self, socks5_config: Socks5) -> Self {
         self.socks5_config = Some(socks5_config);
@@ -377,6 +394,7 @@ where
 
     /// Use an externally managed bandwidth controller instead of having the client spin up its own.
     /// only for advanced use
+    #[cfg(feature = "bandwidth")]
     #[must_use]
     pub fn with_custom_bandwidth_provider(
         mut self,
@@ -395,15 +413,17 @@ where
     /// Construct a [`DisconnectedMixnetClient`] from the setup specified.
     #[allow(clippy::result_large_err)]
     pub fn build(self) -> Result<DisconnectedMixnetClient<S>> {
-        let mut client = DisconnectedMixnetClient::new(
-            self.config,
-            self.socks5_config,
-            self.storage,
-            self.event_tx,
-        )?;
+        let mut client = DisconnectedMixnetClient::new(self.config, self.storage, self.event_tx)?;
 
+        #[cfg(feature = "socks5")]
+        {
+            client.socks5_config = self.socks5_config;
+        }
         client.custom_gateway_transceiver = self.custom_gateway_transceiver;
-        client.custom_bandwidth_provider = self.custom_bandwidth_provider;
+        #[cfg(feature = "bandwidth")]
+        {
+            client.custom_bandwidth_provider = self.custom_bandwidth_provider;
+        }
         client.custom_topology_provider = self.custom_topology_provider;
         client.custom_shutdown = self.custom_shutdown;
         client.wait_for_gateway = self.wait_for_gateway;
@@ -438,6 +458,7 @@ where
     config: Config,
 
     /// Socks5 configuration
+    #[cfg(feature = "socks5")]
     socks5_config: Option<Socks5>,
 
     /// The client can be in one of multiple states, depending on how it is created and if it's
@@ -458,6 +479,7 @@ where
     custom_gateway_transceiver: Option<Box<dyn GatewayTransceiver + Send + Sync>>,
 
     /// advanced usage of an externally managed bandwidth controller
+    #[cfg(feature = "bandwidth")]
     custom_bandwidth_provider: Option<Box<dyn BandwidthTicketProvider>>,
 
     /// Attempt to wait for the selected gateway (if applicable) to come online if it's currently not bonded.
@@ -508,7 +530,6 @@ where
     #[allow(clippy::result_large_err)]
     fn new(
         config: Config,
-        socks5_config: Option<Socks5>,
         storage: S,
         event_tx: Option<EventSender>,
     ) -> Result<DisconnectedMixnetClient<S>> {
@@ -530,12 +551,14 @@ where
 
         Ok(DisconnectedMixnetClient {
             config,
-            socks5_config,
+            #[cfg(feature = "socks5")]
+            socks5_config: None,
             state: BuilderState::New,
             dkg_query_client,
             storage,
             custom_topology_provider: None,
             custom_gateway_transceiver: None,
+            #[cfg(feature = "bandwidth")]
             custom_bandwidth_provider: None,
             wait_for_gateway: false,
             wait_for_initial_topology: false,
@@ -769,6 +792,7 @@ where
 
     /// Creates an associated [`BandwidthAcquireClient`] that can be used to acquire bandwidth
     /// credentials of particular type for this client to consume.
+    #[cfg(feature = "bandwidth")]
     pub async fn create_bandwidth_client(
         &self,
         mnemonic: String,
@@ -801,6 +825,7 @@ where
         .await
     }
 
+    #[cfg(feature = "bandwidth")]
     pub fn begin_bandwidth_import(&self) -> BandwidthImporter<'_, S::CredentialStore> {
         BandwidthImporter::new(self.storage.credential_store())
     }
@@ -852,6 +877,7 @@ where
             base_builder = base_builder.with_gateway_transceiver(gateway_transceiver);
         }
 
+        #[cfg(feature = "bandwidth")]
         if let Some(bandwidth_provider) = self.custom_bandwidth_provider {
             base_builder = base_builder.with_custom_bandwidth_provider(bandwidth_provider);
         }
@@ -892,6 +918,7 @@ where
     ///     let client = client.connect_to_mixnet_via_socks5().await.unwrap();
     /// }
     /// ```
+    #[cfg(feature = "socks5")]
     pub async fn connect_to_mixnet_via_socks5(self) -> Result<Socks5MixnetClient> {
         let socks5_config = self
             .socks5_config
@@ -945,6 +972,7 @@ where
     /// }
     /// ```
     pub async fn connect_to_mixnet(self) -> Result<MixnetClient> {
+        #[cfg(feature = "socks5")]
         if self.socks5_config.is_some() {
             return Err(Error::Socks5Config { set: true });
         }
